+++ docs/docs/resources/changelog.md
@@ -18,6 +18,14 @@ During the alpha period, things might break! We take breaking changes very serio
 - 🚨 BREAKING: Renamed the `classic` theme to `shoelace`
 - 🚨 BREAKING: Renamed `pulse` attribute in `<wa-badge>` to `attention="pulse"` and added `attention="bounce"` [issue:#940]
 - 🚨 BREAKING: Renamed the `vertical` attribute to `orientation="vertical"` in `<wa-split-panel>` and `<wa-divider>` to align with other components and the platform [issue:674]
+- 🚨 BREAKING: Renamed certain boolean attributes to be more consistent using the `with-*` and `without-*` pattern:
+  - `<wa-color-picker no-format-toggle>` => `<wa-color-picker without-format-toggle>`
+  - `<wa-format-number no-grouping>` => `<wa-format-number without-grouping>`
+  - `<wa-input no-spin-buttons>` => `<wa-input without-spin-buttons>`
+  - `<wa-input clearable>` => `<wa-input with-clear>`
+  - `<wa-select clearable>` => `<wa-select with-clear>`
+  - `<wa-tab-group no-scroll-controls>` => `<wa-tab-group without-scroll-controls>`
+  - `<wa-tag removable>` => `<wa-tag with-remove>`
 - Added a `min-block-size` to `<wa-divider orientation="vertical">` to ensure the divider is visible regardless of container height [issue:675]
 - Fixed a bug in `<wa-radio-group>` that caused radios to uncheck when assigning a numeric value [issue:924]
 - Fixed `<wa-button-group>` so dividers properly show between buttons